requirements.txt
rich
psutils
py-cpuinfo